Speed-up stopping a service
Don't wait for the next step before terminating. As best effort signal the delay semaphore to wake up early. Stores might not be atomic and we signal the last delays semaphore but that is fine and we just have to wait the full delay. This is mostly about fast shutdown during tests